pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "truenorth"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alpaca-py",
    "anthropic",
    "fastapi",
    "langfuse",
    "massive",
    "ollama",
    "psycopg[binary]",
    "pydantic-settings",
    "pytz",  # transitive dep of alpaca-py, not declared by it
    "pyyaml",
    "uvicorn[standard]",
    "yfinance",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "pyright"]

[project.scripts]
truenorth = "truenorth.main:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "I",  # isort
    "F",  # pyflakes
]

[tool.hatch.build.targets.wheel]
packages = ["truenorth"]
